Начальная реализация обработки символьных ссылок

+++ kernel/fs/vfs/inode.h
@@ -22,7 +22,7 @@ struct inode_operations {
     int      (*rmdir)(struct inode*, struct dentry*);
     int      (*mknod)(struct inode*, const char*, mode_t);
     int      (*symlink)(struct inode*, const char*, const char*);
-    int      (*readlink)(struct inode*, char*, size_t);
+    ssize_t  (*readlink)(struct inode*, char*, size_t);
 };
 
 struct file_operations;
